+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 * sdk: remove `nwc` dep ([Yuki Kishimoto])
 * database: remove `lru`, `thiserror` and `tracing` deps ([Yuki Kishimoto])
 * lmdb: remove `thiserror` and `tracing` deps ([Yuki Kishimoto])
+* indexeddb: remove `thiserror` and `tracing` deps ([Yuki Kishimoto])
 * zapper: remove `thiserror` dep ([Yuki Kishimoto])
 * nwc: remove `thiserror` dep and unnecessary `Error::Zapper` variant ([Yuki Kishimoto])
 * ffi: drop support for `i686-linux-android` target ([Yuki Kishimoto])